make sorting of benchmarks independent of presentation style
The function argument to selectBenchmarks to generate the benchmarks now takes
one more argument of type `GroupStyle`. It is useful when we want to sort the
benchmarks in a way independent of the actual presentation style. For example
we may want to sort the benchmarks using `PercentDiff` but the presentation
style in the config could be `Absolute` so that we show aboslute values in the
report.